Update 03-styles instruction text
Updated the style instructions to clarify where the style classname comes from (e.g. it is created with the prefix of `is-style-` plus whatever the name of the style is). 

**Motivation**
When I was going through this training I was very confused at this part. The instructions seemed to indicate I was supposed to provide the class name instead of it being inferred by the block style config. 

I feel like the above is a small way to make the instructions a bit more obvious and maybe help someone else who gets confused by this.